fix(mcp): resolve compact results CI and review feedback
